projet_kinect/requirements.txt
# Dependances Python 3 pour le sous-projet Kinect
python-dotenv>=1.0.0
streamlit>=1.30.0
matplotlib>=3.7.0
numpy>=1.24.0
opencv-python>=4.8.0
mediapipe>=0.10.0
